[#120] fix: request amount 데이터 타입을 BigDecimal 로 변경

+++ src/main/java/com/modeunsa/boundedcontext/payment/app/dto/PaymentAccountDepositResponse.java
@@ -5,9 +5,9 @@ import lombok.Getter;
 
 @Getter
 public class PaymentAccountDepositResponse {
-  private final long balance;
+  private final BigDecimal balance;
 
   public PaymentAccountDepositResponse(BigDecimal balance) {
-    this.balance = balance.longValue();
+    this.balance = balance;
   }
 }
